refactor: create contants in yandex map and some changes
